requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
idna==2.10
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.0
nltk==3.5
numpy==1.19.1
oauthlib==3.1.0
Pillow==7.2.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tqdm==4.48.0
tweepy==3.9.0
urllib3==1.25.10
wordcloud==1.7.0
